made writing and reading global
